Uploading initial files for demo
The engine uses a binary file format for all maps, and uses png images for graphics. I made all the resources here myself. Work on a map editor is in progress, but not yet complete, so maps have to be made by hand currently if you want to experiment with it. I'll upload an early prototype for the map editor later. Scroll different display modes using the tab key. Exit the program with the escape key. All other controls use the arrow keys currently.